refactor: move e2e to test folder
