+++ .eslintrc
@@ -39,13 +39,11 @@
 		},
 		{
 			"files": [
-				"test/**",
-				"test-e2e/**"
+				"test/**"
 			],
 			"rules": {
 				"@typescript-eslint/no-explicit-any": "off",
 				"@typescript-eslint/no-non-null-assertion": "off",
-				"no-unused-expressions": "off",
 				"no-restricted-properties": [
 					"error",
 					{
