+++ test/tests/e2e/cases/adopted-probes.test.ts
@@ -1,7 +1,7 @@
 import got from 'got';
 import { expect } from 'chai';
-import { client } from '../../src/lib/sql/client.js';
-import { ADOPTED_PROBES_TABLE } from '../../src/lib/adopted-probes.js';
+import { client } from '../../../../src/lib/sql/client.js';
+import { ADOPTED_PROBES_TABLE } from '../../../../src/lib/adopted-probes.js';
 import { waitProbeInCity } from '../utils.js';
 
 describe('adopted probe', () => {
