+++ test/tests/e2e/cases/dns.test.ts
@@ -30,7 +30,7 @@ describe('dns mesurement', () => {
 
 		expect(response.body.status).to.equal('finished');
 		expect(response.body.results[0].result.status).to.equal('finished');
-		expect(response.body.results[0].result.hops.length > 0).to.be.true;
+		expect(response.body.results[0].result.hops.length).to.be.above(0);
 		expect(response).to.matchApiSchema();
 	});
 });
