+++ test/tests/e2e/cases/location.test.ts
