+++ test/tests/e2e/cases/ping.test.ts
