+++ wallaby.e2e.js
@@ -6,20 +6,20 @@ export default function wallaby () {
 			'public/v1/*',
 			'public/**/*.yaml',
 			'test/plugins/**/*',
-			'test-e2e/setup-probe.ts',
-			'test-e2e/utils.ts',
-			'test-e2e/docker.ts',
+			'test/tests/e2e/setup-probe.ts',
+			'test/tests/e2e/utils.ts',
+			'test/tests/e2e/docker.ts',
 			'src/**/*.ts',
 			'knexfile.js',
 			'package.json',
 		],
 		tests: [
-			'test-e2e/**/*.test.ts',
+			'test/tests/e2e/**/*.test.ts',
 		],
 
 		setup (w) {
 			const path = require('path');
-			w.testFramework.addFile(path.resolve(process.cwd(), 'test-e2e/setup-probe.js'));
+			w.testFramework.addFile(path.resolve(process.cwd(), 'test/tests/e2e/setup-probe.js'));
 			w.testFramework.timeout(20000);
 		},
